Let's try deploying without database
