+++ heroku.yml
@@ -1,4 +1,5 @@
 build:
   docker:
     web: Dockerfile.backend.production
-    mongodb: Dockerfile.mongodb
+  config:
+    PORT: 1337
